@@ -1,91 +1,390 @@
 {
-  "name": "luminati-proxy",
-  "version": "0.8.6",
+  "author": "Luminati Networks Ltd <lpm@luminati.io> (https://luminati.io/)",
+  "bin": {
+    "luminati": "bin/index.js",
+    "luminati-proxy": "bin/index.js"
+  },
+  "config": {
+    "unsafe-perm": true
+  },
+  "bugs": {
+    "email": "lpm@luminati.io"
+  },
+  "build": {
+    "appId": "luminati.proxy_manager",
+    "asar": false,
+    "artifactName": "${name}-${version}.${ext}",
+    "directories": {
+      "output": "d"
+    },
+    "files": [
+      "!publish*.js*",
+      "!**/.jakecache${/*}",
+      "!**/test_ln",
+      "!**/xvfs_ln",
+      "!**/builddir_ln",
+      "!build.lock",
+      "!luminati-proxy-*",
+      "!latest.yml",
+      "!deploy${/*}"
+    ],
+    "npmRebuild": false,
+    "nsis": {
+      "artifactName": "${name}-${version}-setup.${ext}",
+      "warningsAsErrors": false
+    },
+    "productName": "Luminati Proxy Manager",
+    "publish": {
+      "provider": "generic",
+      "url": "http://luminati.io/static/lpm"
+    },
+    "win": {
+      "target": "nsis"
+    },
+    "mac": {
+      "target": "tar.gz"
+    },
+    "dmg": {
+      "contents": [
+        {
+          "x": 110,
+          "y": 150
+        },
+        {
+          "x": 240,
+          "y": 150,
+          "type": "link",
+          "path": "/Applications"
+        }
+      ]
+    }
+  },
+  "dependencies": {
+    "body-parser": "1.15.2",
+    "check-node-version": "2.0.1",
+    "country-language": "0.1.7",
+    "electron-updater": "2.8.9",
+    "express": "4.14.0",
+    "font-awesome": "4.7.0",
+    "flag-icon-css": "2.8.0",
+    "http-shutdown": "1.1.0",
+    "json-stable-stringify": "1.0.1",
+    "json2csv": "3.7.3",
+    "lodash": "4.17.2",
+    "node-forge": "0.6.46",
+    "notosans-fontface": "1.0.1",
+    "pm2": "2.10.1",
+    "request": "2.79.0",
+    "request-stats": "2.0.1",
+    "semver": "5.3.0",
+    "@luminati-io/socksv5": "0.0.7",
+    "sqlite3": "4.0.2",
+    "sudo-prompt": "6.2.1",
+    "taskkill": "2.0.0",
+    "tasklist": "3.1.0",
+    "tough-cookie-file-store": "1.2.0",
+    "universal-analytics": "0.4.9",
+    "yargs": "6.5.0",
+    "opn": "5.1.0",
+    "prismjs": "1.8.4",
+    "classnames": "2.2.5",
+    "netmask": "1.0.6",
+    "rotating-file-stream": "1.3.4",
+    "server-destroy": "1.0.1",
+    "uuid": "3.2.1",
+    "compression": "1.7.2",
+    "json-object-stream": "0.1.1",
+    "ws": "5.1.1",
+    "diacritic": "0.0.2",
+    "underscore": "1.8.3",
+    "node-getopt": "0.2.3",
+    "rimraf": "2.5.2",
+    "log-buffer": "0.0.3",
+    "cheerio": "1.0.0-rc.2",
+    "puppeteer": "1.8.0",
+    "nightmare": "3.0.1",
+    "prompt-sync": "4.1.6",
+    "lokijs": "1.5.5",
+    "ps-list": "5.0.0"
+  },
   "description": "A configurable local proxy for luminati.io",
-  "homepage": "https://luminati.io/",
-  "main": "index.js",
+  "devDependencies": {
+    "animate.css": "3.5.2",
+    "babel-cli": "6.24.1",
+    "babel-loader": "7.1.1",
+    "babel-plugin-transform-class-properties": "6.24.1",
+    "babel-plugin-transform-react-jsx": "6.24.1",
+    "babel-preset-env": "1.6.1",
+    "babel-preset-react": "6.24.1",
+    "babel-plugin-transform-object-rest-spread": "6.23.0",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "bootstrap": "3.3.7",
+    "codemirror": "5.21.0",
+    "css-loader": "0.28.4",
+    "electron": "1.7.1",
+    "electron-builder": "16.6.1",
+    "electron-rebuild": "1.7.2",
+    "es6-shim": "0.35.3",
+    "file-loader": "0.11.2",
+    "html-webpack-plugin": "2.30.1",
+    "jquery": "3.1.1",
+    "less": "2.7.2",
+    "less-loader": "4.0.5",
+    "lolex": "2.1.1",
+    "mocha": "3.2.0",
+    "moment": "2.17.1",
+    "nock": "9.0.11",
+    "react": "16.3.2",
+    "react-bootstrap": "0.31.0",
+    "react-bootstrap-typeahead": "2.0.2",
+    "react-dom": "16.1.1",
+    "react-router-dom": "4.2.2",
+    "style-loader": "0.18.2",
+    "swagger-ui": "2.2.8",
+    "url-loader": "0.5.9",
+    "file-saver": "1.3.3",
+    "webpack": "3.4.0",
+    "recompose": "0.26.0",
+    "prop-types": "15.6.0",
+    "react-waypoint": "8.0.1",
+    "react-select": "2.0.0"
+  },
   "engines": {
-    "node": ">=4&&<7"
+    "node": ">=10.0",
+    "npm": "4.6.1"
   },
-  "recomendedNode": "6",
-  "scripts": {
-    "start": "bin/luminati.js",
-    "test": "./node_modules/mocha/bin/mocha"
-  },
+  "homepage": "https://luminati.io/",
   "keywords": [
     "luminati",
     "proxy",
     "hola"
   ],
-  "author": "Gilad Novik <gilad@hola.org>",
-  "license": "MIT",
-  "preferGlobal": "true",
-  "bin": {
-    "luminati-proxy": "bin/luminati.js",
-    "luminati": "bin/luminati.js"
+  "license": "ISC",
+  "lpm": {
+    "files": [
+      "bin/ca.crt",
+      "bin/ca.key",
+      "bin/check_compat.js",
+      "bin/index.js",
+      "bin/lum_common.js",
+      "bin/lum_electron.js",
+      "bin/lum_electron_index.js",
+      "bin/lum_node.js",
+      "bin/lum_node_index.js",
+      "bin/pub/favicon.ico",
+      "bin/pub/font/lato_bold.woff2",
+      "bin/pub/font/lato_regular.woff2",
+      "bin/pub/img/add.svg",
+      "bin/pub/img/arrow_down.svg",
+      "bin/pub/img/attention.svg",
+      "bin/pub/img/bell.svg",
+      "bin/pub/img/check.svg",
+      "bin/pub/img/checker.png",
+      "bin/pub/img/check_radio.svg",
+      "bin/pub/img/columns.svg",
+      "bin/pub/img/config.svg",
+      "bin/pub/img/country.png",
+      "bin/pub/img/create.svg",
+      "bin/pub/img/c_download.svg",
+      "bin/pub/img/c_download_active.svg",
+      "bin/pub/img/c_edit.svg",
+      "bin/pub/img/c_edit_active.svg",
+      "bin/pub/img/debug.svg",
+      "bin/pub/img/delete.svg",
+      "bin/pub/img/down.svg",
+      "bin/pub/img/duplicate.svg",
+      "bin/pub/img/edit.svg",
+      "bin/pub/img/error.svg",
+      "bin/pub/img/errors.svg",
+      "bin/pub/img/faq.svg",
+      "bin/pub/img/flag_any_country.svg",
+      "bin/pub/img/general.svg",
+      "bin/pub/img/google.svg",
+      "bin/pub/img/history.svg",
+      "bin/pub/img/howto.svg",
+      "bin/pub/img/icon_chrome.jpg",
+      "bin/pub/img/icon_explorer.png",
+      "bin/pub/img/icon_firefox.jpg",
+      "bin/pub/img/icon_safari.jpg",
+      "bin/pub/img/ic_browser.svg",
+      "bin/pub/img/ic_cancel.svg",
+      "bin/pub/img/ic_checkmark.svg",
+      "bin/pub/img/ic_code.svg",
+      "bin/pub/img/ic_columns.svg",
+      "bin/pub/img/ic_delete.svg",
+      "bin/pub/img/ic_download_mini.svg",
+      "bin/pub/img/ic_dup.svg",
+      "bin/pub/img/ic_edit.svg",
+      "bin/pub/img/ic_edit_disabled.svg",
+      "bin/pub/img/ic_history.svg",
+      "bin/pub/img/ic_history_disabled.svg",
+      "bin/pub/img/ic_off.svg",
+      "bin/pub/img/ic_plus.svg",
+      "bin/pub/img/ic_refresh.svg",
+      "bin/pub/img/ic_refresh_disabled.svg",
+      "bin/pub/img/ic_refresh_mini.svg",
+      "bin/pub/img/ic_status_details_error.svg",
+      "bin/pub/img/ic_status_details_warning.svg",
+      "bin/pub/img/ic_status_ok.svg",
+      "bin/pub/img/ic_status_ok_with_warning.svg",
+      "bin/pub/img/ic_status_red.svg",
+      "bin/pub/img/ic_status_undefined.svg",
+      "bin/pub/img/ic_table.svg",
+      "bin/pub/img/ic_table_disabled.svg",
+      "bin/pub/img/ic_thumbs_down.svg",
+      "bin/pub/img/ic_thumbs_up.svg",
+      "bin/pub/img/ic_warning.svg",
+      "bin/pub/img/info.svg",
+      "bin/pub/img/large_icons.png",
+      "bin/pub/img/loader.gif",
+      "bin/pub/img/logo-80legs.svg",
+      "bin/pub/img/logo-crawlera.svg",
+      "bin/pub/img/logo-import.svg",
+      "bin/pub/img/logo.png",
+      "bin/pub/img/logo_1.svg",
+      "bin/pub/img/logo_big.svg",
+      "bin/pub/img/logs.svg",
+      "bin/pub/img/lpm_infographics.png",
+      "bin/pub/img/luminati_logo_2.svg",
+      "bin/pub/img/lum_logo_short.svg",
+      "bin/pub/img/medium_icons.png",
+      "bin/pub/img/ok_warning.svg",
+      "bin/pub/img/pool.svg",
+      "bin/pub/img/preset_icon.png",
+      "bin/pub/img/proxies.svg",
+      "bin/pub/img/proxy_ext.svg",
+      "bin/pub/img/proxy_lum.svg",
+      "bin/pub/img/refresh.svg",
+      "bin/pub/img/request_headers.svg",
+      "bin/pub/img/resource_document_icon_small.png",
+      "bin/pub/img/rotation.svg",
+      "bin/pub/img/rules.svg",
+      "bin/pub/img/setup_guide.svg",
+      "bin/pub/img/small_icons.png",
+      "bin/pub/img/social_btns.svg",
+      "bin/pub/img/speed.svg",
+      "bin/pub/img/stats.svg",
+      "bin/pub/img/step_create_active.svg",
+      "bin/pub/img/step_create_disabled.svg",
+      "bin/pub/img/step_done.svg",
+      "bin/pub/img/step_not_done.svg",
+      "bin/pub/img/step_req_active.svg",
+      "bin/pub/img/step_req_disabled.svg",
+      "bin/pub/img/step_test_active.svg",
+      "bin/pub/img/step_test_disabled.svg",
+      "bin/pub/img/success.svg",
+      "bin/pub/img/target.svg",
+      "bin/pub/img/tester.svg",
+      "bin/pub/img/tools.svg",
+      "bin/pub/img/trash.svg",
+      "bin/pub/img/treeoutline_triangles.png",
+      "bin/pub/img/up.svg",
+      "bin/pub/img/warning.svg",
+      "bin/pub/img/zones.svg",
+      "bin/pub/img/zone_icon.png",
+      "changelog.json",
+      "lib/cities.js",
+      "lib/context.js",
+      "lib/db_history.json",
+      "lib/find_iface.js",
+      "lib/hash.js",
+      "lib/hutil-loader.js",
+      "lib/locations.js",
+      "lib/log.js",
+      "lib/loki.js",
+      "lib/luminati.js",
+      "lib/manager.js",
+      "lib/md5_util.js",
+      "lib/perr.js",
+      "lib/req_process.js",
+      "lib/router.js",
+      "lib/rules.js",
+      "lib/session.js",
+      "lib/socks.js",
+      "lib/ssl.js",
+      "lib/timeline.js",
+      "lib/tracer.js",
+      "lib/username.js",
+      "lib/util.js",
+      "src/pub/app.js",
+      "src/pub/app.less",
+      "src/pub/chrome_widgets.js",
+      "src/pub/common.js",
+      "src/pub/config.js",
+      "src/pub/har_preview.js",
+      "src/pub/har_viewer.js",
+      "src/pub/howto.js",
+      "src/pub/index.html",
+      "src/pub/instructions.js",
+      "src/pub/json_viewer.js",
+      "src/pub/login.js",
+      "src/pub/logs.js",
+      "src/pub/nav.js",
+      "src/pub/notif_center.js",
+      "src/pub/overview.js",
+      "src/pub/proxies.js",
+      "src/pub/proxy_add.js",
+      "src/pub/proxy_blank.js",
+      "src/pub/proxy_edit/common.js",
+      "src/pub/proxy_edit/debug.js",
+      "src/pub/proxy_edit/fields.js",
+      "src/pub/proxy_edit/general.js",
+      "src/pub/proxy_edit/headers.js",
+      "src/pub/proxy_edit/index.js",
+      "src/pub/proxy_edit/rotation.js",
+      "src/pub/proxy_edit/rules.js",
+      "src/pub/proxy_edit/speed.js",
+      "src/pub/proxy_edit/targeting.js",
+      "src/pub/proxy_tester.js",
+      "src/pub/report_bug.js",
+      "src/pub/schema.js",
+      "src/pub/settings.js",
+      "src/pub/stats.js",
+      "src/pub/tracer.js",
+      "src/pub/util.js",
+      "src/pub/whitelist_ips.js",
+      "test.js",
+      "util/config.js",
+      "util/lpm_config.js",
+      "util/lpm_file.js",
+      "util/lpm_util.js",
+      "webpack.config.js"
+    ],
+    "md5": "6a84fb15b41f2930a1cce70b506d3d0d"
   },
+  "main": "bin/index.js",
+  "name": "@luminati-io/luminati-proxy",
+  "contributors": [
+    {
+      "name": "Colin Shribman",
+      "email": "colin@luminati.io"
+    },
+    {
+      "name": "Assaf Zarfaty",
+      "email": "assafz@luminati.io"
+    },
+    {
+      "name": "Krzysztof Fudali",
+      "email": "krzysztof@luminati.io"
+    }
+  ],
+  "optionalDependencies": {
+    "heapdump": "0.3.9",
+    "posix": "4.1.2",
+    "sinon": "1.17.7"
+  },
+  "preferGlobal": true,
+  "recomendedNode": ">=10.0",
   "repository": {
     "type": "git",
-    "url": "git://github.com/hola/luminati-proxy.git"
+    "url": "git://github.com/luminati-io/luminati-proxy.git"
   },
-  "bugs": {
-    "email": "gilad@hola.org"
+  "scripts": {
+    "electron": "electron .",
+    "start": "bin/index.js",
+    "test": "./node_modules/mocha/bin/mocha",
+    "test-win": "node node_modules/mocha/bin/mocha",
+    "build": "node_modules/webpack/bin/webpack.js"
   },
-  "dependencies": {
-    "angular": "^1.5.8",
-    "angular-chart.js": "^1.0.3",
-    "babel-core": "^6.18.0",
-    "babel-loader": "^6.2.6",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-2": "^6.18.0",
-    "body-parser": "^1.15.2",
-    "bootstrap": "^3.3.7",
-    "bootstrap-datepicker": "^1.6.4",
-    "chart.js": "^2.3.0",
-    "check-node-version": "^1.1.2",
-    "codemirror": "^5.20.2",
-    "country-data": "^0.0.31",
-    "country-language": "^0.1.7",
-    "css-loader": "^0.25.0",
-    "express": "^4.14.0",
-    "file-loader": "^0.9.0",
-    "http-shutdown": "^1.1.0",
-    "hutil": "^0.1.7",
-    "jquery": "^3.1.1",
-    "json-stable-stringify": "^1.0.1",
-    "lodash": "^4.16.4",
-    "memory-fs": "^0.3.0",
-    "moment": "^2.15.2",
-    "netmask": "^1.0.6",
-    "node-forge": "^0.6.44",
-    "notosans-fontface": "^1.0.1",
-    "pm2": "^2.1.0",
-    "react-bootstrap": "^0.30.5",
-    "react-lite": "^0.15.25",
-    "react-redux": "^4.4.5",
-    "react-router": "^3.0.0",
-    "redux": "^3.6.0",
-    "request": "^2.76.0",
-    "request-stats": "^1.0.1",
-    "require-css": "^0.1.8",
-    "requirejs": "^2.3.2",
-    "socket.io": "^1.5.1",
-    "socksv5": "0.0.6",
-    "sqlite3": "^3.1.8",
-    "style-loader": "^0.13.1",
-    "sudo-prompt": "^6.2.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.3",
-    "winston": "^2.2.0",
-    "yargs": "^6.3.0"
-  },
-  "optionalDependencies": {
-    "webshot": "^0.18.0"
-  },
-  "devDependencies": {
-    "mocha": "^3.1.2"
-  }
+  "version": "1.108.399"
 }
