@@ -1,25 +1,25 @@
 // LICENSE_CODE ZON ISC
 'use strict'; /*jslint node:true, esnext:true*/
-const socks = require('socksv5');
+const socks = require('@luminati-io/socksv5');
 const net = require('net');
-const hutil = require('hutil');
+const lutil = require('./util.js');
 const util = require('util');
 const events = require('events');
 const log = require('./log.js');
-const find_iface = require('./find_iface.js');
-const etask = hutil.etask;
+const etask = require('../util/etask.js');
+const zerr = require('../util/zerr.js');
 module.exports = Socks;
 
 function Socks(opt){
     events.EventEmitter.call(this);
-    this._log = log(`${opt.local}-${opt.remote}`, opt.log);
+    this._log = log(`socks:${opt.port}-${opt.port}`, opt.log);
     this.opt = opt;
     this.server = socks.createServer(this.handler.bind(this));
     this.server.useAuth(socks.auth.None());
     this.connections = {};
     this.next_id = 0;
     this.server.on('error', err=>{
-        this._log.error(err);
+        this._log.error('server error %O', err);
         this.emit('error', err);
     });
 }
@@ -27,60 +27,125 @@ function Socks(opt){
 util.inherits(Socks, events.EventEmitter);
 
 Socks.prototype.handler = function(info, accept, deny){
-    let dst, socket = accept(true);
-    const id = this.next_id++;
-    if (info.dstPort==80 || info.dstPort==443)
-    {
-        dst = net.connect(this.opt.remote, '127.0.0.1');
-        if (info.dstPort==443)
-        {
-            this._log.debug('Socks https connection', info);
-            dst.on('connect', ()=>{
-                dst.write(util.format('CONNECT %s:%d HTTP/1.1\r\n'+
-                'Host: %s:%d\r\n\r\n', info.dstAddr, info.dstPort,
-                info.dstAddr, info.dstPort));
-                socket.pipe(dst);
-            }).once('data', ()=>{ dst.pipe(socket); });
-        }
-        else
-            this._log.debug('Socks http connection', info);
-    }
-    else
-    {
-        this._log.debug('Socks direct connection', info);
-        dst = net.connect(info.dstPort, info.dstAddr);
-    }
-    socket._connectionId = dst._connectionId = id;
-    this.connections[id] = {socket, dst};
-    const end = ()=>delete this.connections[id];
-    socket.on('end', end);
-    dst.on('end', end);
-    socket.on('error', error=>{
-        this._log.warn('Socks connection error', {error, socket});
-    });
-    dst.on('error', error=>{
-        this._log.warn('Socks forward connection error', {error, dst});
-    });
-    if (info.dstPort!=443)
-        socket.pipe(dst).pipe(socket);
+    if (info.dstPort==80)
+        return this.proxy_http(info, accept, deny);
+    // XXX krzysztof: this is quick fix. solution should check zon's config
+    // to see what ports are allowed
+    if (info.dstPort==443||info.dstPort==43||info.dstPort==25)
+        return this.proxy_https(info, accept, deny);
+    this.bypass(info, accept, deny);
 };
 
-Socks.prototype.start = etask._fn(function*start(_this){
-    let args = [_this.opt.local, find_iface(_this.opt.iface || '0.0.0.0')];
-    _this._log.debug('Socks starting:', args);
-    yield etask.cb_apply(_this.server, '.listen', args);
-    _this.port = _this.server.address().port;
+Socks.prototype.get_src_header = function(info){
+    return 'x-lpm-src-addr: '+info.srcAddr.replace('::ffff:', '');
+};
+
+Socks.prototype.lpm_connect = function(){
+    return net.connect(this.server._srv.address().port, '127.0.0.1');
+};
+
+Socks.prototype.proxy_https = function(info, accept, deny){
+    const dst = this.lpm_connect();
+    this._log.debug('Socks https connection %O', info);
+    dst.on('connect', ()=>{
+        dst.write(
+            `CONNECT ${info.dstAddr}:${info.dstPort} HTTP/1.1\r\n`+
+            `Host: ${info.dstAddr}:${info.dstPort}\r\n`+
+            `${this.get_src_header(info)}\r\n\r\n`);
+
+        dst.once('data', d=>{
+            let res;
+            try {
+                res = lutil.parse_http_res(d.toString('utf8'));
+            } catch(e){
+                this._log.error('could not parse http resp %s', zerr.e2s(e));
+                dst.end();
+                deny();
+                return;
+            }
+            if (res.headers['x-lpm-error'])
+            {
+                this._log.error('connection error %s:%s %s %s: %s',
+                    info.dstAddr, info.dstPort, res.status_code,
+                    res.status_message, res.headers['x-lpm-error']||'');
+            }
+            let socket = accept(true);
+            if (!socket)
+            {
+                this._log.warn('connection closed by client %O', info);
+                dst.end();
+                return;
+            }
+            socket.pipe(dst).pipe(socket);
+            this._handle_connection(socket, dst);
+        });
+    });
+};
+
+Socks.prototype.proxy_http = function(info, accept, deny){
+    const dst = this.lpm_connect();
+    const socket = accept(true);
+    this._log.debug('Socks http connection %O', info);
+    socket.once('data', data=>{
+        let d = data.toString().split('\r\n');
+        d.splice(d.indexOf(''), 0, this.get_src_header(info));
+        dst.write(d.join('\r\n'));
+        socket.pipe(dst).pipe(socket);
+    });
+    return this._handle_connection(socket, dst);
+};
+
+Socks.prototype.bypass = function(info, accept, deny){
+    const socket = accept(true);
+    this._log.debug('Socks direct connection %O', info);
+    let dst = net.connect(info.dstPort, info.dstAddr);
+    socket.pipe(dst).pipe(socket);
+    this._handle_connection(socket, dst);
+};
+
+Socks.prototype._handle_connection = function(socket, dst){
+    const id = this.next_id++;
+    socket._connectionId = dst._connectionId = id;
+    this.connections[id] = {socket, dst};
+    const end = src=>()=>{
+        this._log.debug('connection end %s connid:%s', src, id);
+        delete this.connections[id];
+    };
+    socket.on('end', end('socket'));
+    dst.on('end', end('dst'));
+    socket.on('error', error=>{
+        this._log.error('Socks connection error, %O', error);
+    });
+    dst.on('error', error=>{
+        this._log.error('Socks forward connection error, %O', error);
+    });
+};
+
+Socks.prototype.connect = function(socket){
+    this.server._onConnection(socket);
+};
+
+Socks.prototype.set_tcp_server = etask._fn(
+function*set_tcp_server(_this, tcp_server){
+    const sp = etask(function*close_def_srv(){ return yield this.wait(); });
+    _this.server._srv.close(function(){ sp.continue(); });
+    yield sp;
+    _this.server._srv = tcp_server;
+    _this.server._srv.on('error', err=>_this.server.emit('error', err))
+    .on('listening', ()=>_this.server.emit('listening'))
+    .on('close', ()=>_this.server.emit('close'));
 });
 
-Socks.prototype.stop = etask._fn(function*stop(_this, force){
+Socks.prototype.stop = function(force){
+    this._log.debug('Socks stopping');
     if (force)
     {
-        for (let k in _this.connections)
+        for (let k in this.connections)
         {
-            let c = _this.connections[k];
+            let c = this.connections[k];
             c.socket.destroy();
             c.dst.destroy();
         }
     }
-    yield etask.nfn_apply(_this.server, '.close', []);
-});
+    this._log.debug('Socks stopped');
+};
