@@ -6,7 +6,7 @@ const tls = require('tls');
 const forge = require('node-forge');
 const pki = forge.pki;
 const E = module.exports = ssl;
-const keys = pki.rsa.generateKeyPair(2048);
+let keys;
 E.ca = {
     cert: fs.readFileSync(path.join(__dirname, '../bin/ca.crt')),
     key: fs.readFileSync(path.join(__dirname, '../bin/ca.key')),
@@ -16,9 +16,15 @@ function ssl(opt){
     opt = opt||{};
     if (!opt.keys)
     {
+        if (!keys)
+        {
+            keys = pki.rsa.generateKeyPair(2048);
+            keys.privateKeyPem = pki.privateKeyToPem(keys.privateKey);
+            keys.publicKeyPem = pki.publicKeyToPem(keys.publicKey);
+        }
         opt.keys = {
-            privateKey: pki.privateKeyToPem(keys.privateKey),
-            publicKey: pki.publicKeyToPem(keys.publicKey),
+            privateKey: keys.privateKeyPem,
+            publicKey: keys.publicKeyPem,
         };
     }
     if (!opt.ca)
@@ -34,6 +40,13 @@ function ssl(opt){
         cert.validity.notAfter = new Date(Date.now()+10*365*86400000);
         cert.setSubject([{name: 'commonName', value: name}]);
         cert.setIssuer(pki.certificateFromPem(opt.ca.cert).issuer.attributes);
+        cert.setExtensions([{
+            name: 'subjectAltName',
+            altNames: [{
+                type: 2,
+                value: name
+            }]
+        }]);
         cert.sign(pki.privateKeyFromPem(opt.ca.key), forge.md.sha256.create());
         hosts[name] = tls.createSecureContext({
             key: opt.keys.privateKey,
